initial test of surface Stokes and surface Laplacian in serial
