+++ src/CellData/CellData.jl
@@ -26,6 +26,7 @@ export InvMetricCellField
 export MeasureCellField
 export AmbientMapCellField
 export LatLonMapCellField
+export PerpMetricCellField
 
 export Δs
 export vecΔs
@@ -35,6 +36,8 @@ export divs
 export skew_∇s
 export skew_divs
 export dagger
+export vecΔs_2D
+export ∇s_contra
 
 export pushforward_normal
 export pushforward_reference_normal
