+++ src/CellData/SurfaceDiffOps.jl
@@ -167,7 +167,7 @@ function _skew_∇s_no_ad(f, Ω_atlas)
   meas_cf = MeasureCellField(Ω_atlas)
   J_cf = transpose∘∇(ambient_map_cf)
   grad_f_cf = (∇(f)∘ambient_map_cf)⋅J_cf
-  skew_grad_parametric = J_cf⋅(perp∘grad_f_cf)*(1.0/meas_cf) 
+  skew_grad_parametric = J_cf⋅(perp∘grad_f_cf)*(1.0/meas_cf)
 end
 
 function _skew_∇s_ad(f, Ω_atlas)
@@ -215,7 +215,7 @@ end
 
 function _divs_no_ad(f, Ω_atlas)
     # 1/m * div( m * (J^†⋅(f∘ϕ)) ), where J^†=inv(g)⋅Jᵀ
-    # grad(m)⋅(J^†⋅(f∘ϕ)) + 
+    # grad(m)⋅(J^†⋅(f∘ϕ)) +
     # div((J^†⋅(f∘ϕ))) = tr(grad(J^†):(f∘ϕ)) + tr(J^†⋅grad(f∘ϕ))
     # grad(J^†) = grad(inv(g)⋅Jᵀ) = grad(inv(g))⋅Jᵀ + inv(g)⊙grad(Jᵀ)
     metric_cf = MetricCellField(Ω_atlas)
@@ -226,11 +226,11 @@ function _divs_no_ad(f, Ω_atlas)
     f_cf = f∘ambient_map_cf
     grad_f_cf = ∇(f)∘ambient_map_cf
     Jt_cf = ∇(ambient_map_cf)
-    
+
     # grad(inv(g))⋅Jᵀ
     grad_inv_metric_cf = ∇(inv_metric_cf)
     trace_1=Operation(tr)((grad_inv_metric_cf⋅Jt_cf)⋅f_cf)
-    
+
     # inv(g)⋅grad(Jᵀ)
     trace_2 = Operation(tr)((inv_metric_cf ⋅ ∇(Jt_cf))⋅f_cf)
 
@@ -240,20 +240,20 @@ function _divs_no_ad(f, Ω_atlas)
 
     grad_meas_cf = (deriv_sqrt∘det∘metric_cf)*
                    Operation(cpAB)(deriv_det∘metric_cf,gradient(metric_cf))
-   
 
-    return (1.0/meas_cf)*(meas_cf*(trace_1+trace_2+trace_3) + 
+
+    return (1.0/meas_cf)*(meas_cf*(trace_1+trace_2+trace_3) +
                            grad_meas_cf⋅(inv_metric_cf⋅Jt_cf⋅f_cf))
 end
 
 function _skew_divs_no_ad(f, Ω_atlas)
     # -1/m * div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) ), where J^†=inv(g)⋅Jᵀ
     # div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) )
-    # div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) ) = 
+    # div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) ) =
     #   grad(m^2)⋅(inv(g) R(J^†⋅(f∘ϕ))) + m^2 * div(inv(g) R(J^†⋅(f∘ϕ)))
     # div(inv(g) R(J^†⋅(f∘ϕ))) = tr(grad(inv(g))⋅R(J^†⋅(f∘ϕ))) + tr(inv(g)⋅grad(R(J^†⋅(f∘ϕ))))
     #
-    Gridap.Helpers.@notimplemented "skew_divs without automatic differentiation is not implemented yet"  
+    Gridap.Helpers.@notimplemented "skew_divs without automatic differentiation is not implemented yet"
 end
 
 function _skew_divs_ad(f, Ω_atlas)
@@ -263,8 +263,8 @@ function _skew_divs_ad(f, Ω_atlas)
   CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
 end
 
-# Surface divergence of an ambient vector-valued function which is 
-# pulled back using the pseudo-inverse of the jacobian of the ambient 
+# Surface divergence of an ambient vector-valued function which is
+# pulled back using the pseudo-inverse of the jacobian of the ambient
 # map without multiplying by the measure
 function divs(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
               use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
@@ -387,10 +387,10 @@ function _vecΔs_ad(f, Ω_atlas)
 end
 
 function _vecΔs_no_ad(f, Ω_atlas)
-  Gridap.Helpers.@notimplemented "vecΔs without automatic differentiation is not implemented yet"  
+  Gridap.Helpers.@notimplemented "vecΔs without automatic differentiation is not implemented yet"
 end
 
-# Returns the co-vector components of the vector surface laplacian applied to the 
+# Returns the co-vector components of the vector surface laplacian applied to the
 # ambient vector-valued function f
 function vecΔs(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
                 use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
@@ -422,10 +422,10 @@ function _curls_ad(f, Ω_atlas)
 end
 
 function _curls_no_ad(f, Ω_atlas)
-  Gridap.Helpers.@notimplemented "curls without automatic differentiation is not implemented yet"  
+  Gridap.Helpers.@notimplemented "curls without automatic differentiation is not implemented yet"
 end
 
-# Returns the co-vector components of the surface curl operator applied to the 
+# Returns the co-vector components of the surface curl operator applied to the
 # ambient vector-valued function f
 function curls(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
                 use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
@@ -436,4 +436,93 @@ function curls(f::Function, Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMIM{Dc,
                 use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
   curls_trian = use_automatic_differentiation ? _curls_ad(f, Ω_atlas.trian) : _curls_no_ad(f, Ω_atlas.trian)
   Gridap.CellData.GenericCellField(get_data(curls_trian), Ω_atlas, Gridap.CellData.DomainStyle(curls_trian))
+end
+
+
+
+
+
+################################################################################
+## The below operators are required for surface Stokes. Some are repeated above
+## Need to decide how to gather these properly
+################################################################################
+# Contravariant components of the ∇^perp (∇^perp ⋅ ̃u)
+# = J*    1/√g R gradient ( -1/√g div(R g u)     )
+# where u is contravariant component of ̃u, Recall Rg = (√g)^2 g^{-1} R
+_my_skew_sdiv(f, m) = αβ -> detg(m,αβ)*inv_metric(m,αβ)⋅(perp(contra_v(f,m)(αβ)))
+my_skew_surfdiv(f, m) = αβ ->  -1.0/sqrtg(m,αβ)*(divergence(_my_skew_sdiv(f,m))(αβ))
+curly_curl(u, m) = x -> 1/sqrtg(m)(x)*perp(gradient(my_skew_surfdiv(u,m))(x))
+
+
+# Contravariant component of ∇(∇⋅ ̃u)
+# = J *    g^{-1} gradient ( 1/√g div( √g u)   )
+# where u is contravariant component of ̃u
+_my_divs(u,m) = x -> sqrtg(m)(x)*contra_v(u,m)(x)
+my_divs(u, m) = x -> 1/sqrtg(m)(x)*(divergence(_my_divs(u,m))(x))
+
+my_grads_divs(u, m) = x-> gradient(my_divs(u,m))(x)
+contra_grads_divs(u, m) = x-> inv_metric(m,x)⋅my_grads_divs(u,m)(x)
+
+# Contravariant component of the surface vector laplacian in 2D
+vec_laps_2D(u,m) = x ->  contra_grads_divs(u,m)(x) -1.0*curly_curl(u,m)(x)
+
+function _vecΔs_ad_2D(f, Ω_atlas)
+  ambient_map_cf = AmbientMapCellField(Ω_atlas)
+  ambient_maps = Gridap.CellData.get_data(ambient_map_cf)
+  cell_field = lazy_map(m->GenericField(vec_laps_2D(_fm(f,m),m)),ambient_maps)
+  CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
+end
+
+function _vecΔs_no_ad_2D(f, Ω_atlas)
+  Gridap.Helpers.@notimplemented "vecΔs_2D without automatic differentiation is not implemented yet"
+end
+
+# Returns the contravariant components of the vector surface laplacian applied to the
+# ambient vector-valued function f in 2D
+function vecΔs_2D(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
+                use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
+  use_automatic_differentiation ? _vecΔs_ad_2D(f, Ω_atlas) : _vecΔs_no_ad_2D(f, Ω_atlas)
+end
+
+
+function vecΔs_2D(f::Function,
+              Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMIM{Dc,Dc,Da,G,A,P,C,O}};
+              use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
+  vecΔs_2D_trian = use_automatic_differentiation ? _vecΔs_ad_2D(f, Ω_atlas) : _vecΔs_no_ad_2D(f, Ω_atlas)
+  Gridap.CellData.GenericCellField(get_data(vecΔs_2D_trian), Ω_atlas, Gridap.CellData.DomainStyle(vecΔs_2D_trian))
+end
+
+
+# Return the contravariant components of the surface gradient
+# = J *    g^{-1} gradient(f)
+sgrad_contra(f::Function,m::Field) = αβ ->  (inv_metric(m,αβ) ⋅ gradient(f(m))(αβ) )
+sgrad_contra(f::Function) = m -> sgrad_contra(f,m)
+
+function _contra_∇s_no_ad(f, Ω_atlas)
+  ambient_map_cf = AmbientMapCellField(Ω_atlas)
+  inv_metric_cf = InvMetricCellField(Ω_atlas)
+  covariant_basis_cf = transpose∘∇(ambient_map_cf)
+  gradient_f_cf = (∇(f)∘ambient_map_cf)⋅covariant_basis_cf
+  inv_metric_cf⋅gradient_f_cf
+end
+
+function _contra_∇s_ad(f, Ω_atlas)
+  ambient_map_cf = AmbientMapCellField(Ω_atlas)
+  ambient_maps = Gridap.CellData.get_data(ambient_map_cf)
+  cell_field = lazy_map(m->GenericField(sgrad_contra(_fm(f,m),m)),ambient_maps)
+  CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
+end
+
+
+function ∇s_contra(f::Function,
+            Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
+            use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
+  use_automatic_differentiation ? _contra_∇s_ad(f, Ω_atlas) : _contra_∇s_no_ad(f, Ω_atlas)
+end
+
+function ∇s_contra(f::Function,
+            Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMIM{Dc,Dc,Da,G,A,P,C,O}};
+            use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
+  ∇s_trian = use_automatic_differentiation ? _contra_∇s_ad(f, Ω_atlas.trian) : _contra_∇s_no_ad(f, Ω_atlas.trian)
+  Gridap.CellData.GenericCellField(get_data(∇s_trian), Ω_atlas, Gridap.CellData.DomainStyle(∇s_trian))
 end
